refactor(cli): extract status diagnostics boundary
